requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
attrs==22.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.153
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.128
    # via aws-cdk-lib
aws-cdk-aws-apigatewayv2-alpha==2.77.0.a0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
aws-cdk-aws-apigatewayv2-integrations-alpha==2.77.0a0
    # via -r requirements.in
aws-cdk-lib==2.77.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
black==23.3.0
    # via -r requirements.in
cattrs==22.2.0
    # via jsii
click==8.1.3
    # via black
constructs==10.2.8
    # via
    #   -r requirements.in
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-lib
exceptiongroup==1.1.1
    # via cattrs
importlib-resources==5.12.0
    # via jsii
jsii==1.80.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-lib
    #   constructs
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
platformdirs==3.5.0
    # via black
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via black
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-aws-apigatewayv2-alpha
    #   aws-cdk-aws-apigatewayv2-integrations-alpha
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.5.0
    # via
    #   black
    #   jsii
zipp==3.15.0
    # via importlib-resources
